refactor(vm): move the warm-slot pool policy into its own module
The pool is two things bolted together: an LRU with refill accounting that
needs nothing but a `Mutex`, and the staging that spawns a jailed VMM per
slot. Only the first is what R3's actor has to reason about, and it was
already generic over the slot type and unit-tested over `SlotPool<u32>`.

`policy/pool.rs` takes `SlotPool`, `PoolInner`, `FillPlan`,
`MAX_POOLED_SNAPSHOTS` and those tests; `pool.rs` keeps `PreparedSlot`,
`prepare_slot`, `stage_slot`, `destroy_slot`, `spawn_pool_refill` and
`shutdown`, and re-exports `SlotPool` so no call site moves. `policy.rs`
opens the module the rest of this series fills.

Pure move — 295 insertions / 269 deletions, every body byte-identical; the
two moved types widen from `pub(super)` to `pub(in crate::sandbox)` because
their module is now a level deeper.
